Handle board_id kernel command parameter correctly
The kernel command which includes the board_id, may not be constructed correctly
by the legacy kernels (?).
It is assumed that command parameter is X=Y format for each parameter seperated
by space. However it is observed that for board_id, the string came in as
"board_id=S88047C1:board_vol=1046542" which if we split with delimiter '=', it
would become into three pieces, and application is not informed as per the
current algorithm.
Handle this for board_id only. Kernel code may need to be looked to fix at its
end.